refactor: Aside 파일 구조 재정리(#47) - Aside 폴더 안에 Aside.jsx, Aside.css, hooks 폴더, ChatMenu 폴더, FileMenu 폴더, WritingMenu 폴더 생성 - hooks: useShowExtend.js 옮김 - ChatStuMenu: ChatItem 컴포넌트 옮김 - FileMenu: CategoryManageWrapper 컴포넌트, ModifyFileWrapper 컴포넌트 분리, 파일과 관련된 훅들을 hooks 폴더 안에 옮김 - WritingMenu: CreateNewWriting 컴포넌트, 글쓰기와 괸련된 훅들을 hooks 파일에 옮김
